# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend para la optimización de producción de una panadería usando LINGO.",
        "title": "LINGO Bakery API",
        "contact": {
            "name": "Soporte Técnico"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/products": {
            "get": {
                "description": "Obtiene la lista completa de productos con sus precios y costos",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "Listar productos",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Product"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Registra un nuevo producto en el catálogo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "Crear producto",
                "parameters": [
                    {
                        "description": "Datos del producto",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Product"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Product"
                        }
                    }
                }
            }
        },
        "/products/{id}": {
            "get": {
                "description": "Obtiene detalles de un producto, incluyendo ingredientes y máquinas asignadas",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "Obtener producto",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del producto",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Product"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.Ingredient": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "unit": {
                    "description": "kg, litros, unidades, etc.",
                    "type": "string"
                },
                "unit_cost": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.Machine": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.OperationalResource": {
            "type": "object",
            "properties": {
                "available": {
                    "description": "DISP(R)",
                    "type": "number"
                },
                "cost_per_unit": {
                    "description": "CR(R)",
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "resource_id": {
                    "type": "integer"
                }
            }
        },
...
        "models.Product": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "demand": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "ingredients": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ProductIngredient"
                    }
                },
                "machines": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ProductMachine"
                    }
                },
                "max_batch": {
                    "type": "number"
                },
                "min_batch": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "operational_resources": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ProductOperationalResource"
                    }
                },
                "sale_price": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.ProductIngredient": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "ingredient": {
                    "$ref": "#/definitions/models.Ingredient"
                },
                "ingredient_id": {
                    "type": "integer"
                },
                "product_id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "number"
                }
            }
        },
        "models.ProductMachine": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "machine": {
                    "$ref": "#/definitions/models.Machine"
                },
                "machine_id": {
                    "type": "integer"
                },
                "minutes_per_unit": {
                    "type": "number"
                },
                "product_id": {
                    "type": "integer"
                }
            }
        },
        "models.ProductOperationalResource": {
            "type": "object",
            "properties": {
                "consumption_per_batch": {
                    "description": "CM(I,R)",
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "operational_resource": {
                    "$ref": "#/definitions/models.OperationalResource"
                },
                "operational_resource_id": {
                    "type": "integer"
                },
                "product_id": {
                    "type": "integer"
                }
            }
        },
...
    }
}